fix(app): stabilize i18n provider
